# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type RiskLevel = (typeof RiskLevel)[keyof typeof RiskLevel];

export const RiskLevel = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  CRITICAL: "CRITICAL",
} as const;

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  REPORT_RECEIVED: "REPORT_RECEIVED",
  ANALYZED: "ANALYZED",
  CASE_CREATED: "CASE_CREATED",
  ROUTED_TO_HOSPITAL: "ROUTED_TO_HOSPITAL",
  RECEIVED_BY_HOSPITAL: "RECEIVED_BY_HOSPITAL",
  ASSIGNED_TO_DOCTOR: "ASSIGNED_TO_DOCTOR",
  DIAGNOSIS_IN_PROGRESS: "DIAGNOSIS_IN_PROGRESS",
  COMPLETED: "COMPLETED",
} as const;

export type RecommendedAction =
  (typeof RecommendedAction)[keyof typeof RecommendedAction];

export const RecommendedAction = {
  SELF_CARE: "SELF_CARE",
  SCHEDULE_CONSULTATION: "SCHEDULE_CONSULTATION",
  HOSPITAL_VISIT: "HOSPITAL_VISIT",
  EMERGENCY_RESPONSE: "EMERGENCY_RESPONSE",
} as const;

export type ConsultationStatus =
  (typeof ConsultationStatus)[keyof typeof ConsultationStatus];

export const ConsultationStatus = {
  SCHEDULED: "SCHEDULED",
  ACTIVE: "ACTIVE",
  COMPLETED: "COMPLETED",
  CANCELLED: "CANCELLED",
} as const;

export type TransferStatus =
  (typeof TransferStatus)[keyof typeof TransferStatus];

export const TransferStatus = {
  PENDING: "PENDING",
  ACCEPTED: "ACCEPTED",
  IN_TRANSIT: "IN_TRANSIT",
  ARRIVED: "ARRIVED",
  REJECTED: "REJECTED",
  CANCELLED: "CANCELLED",
} as const;

export type Specialty = (typeof Specialty)[keyof typeof Specialty];

export const Specialty = {
  GENERAL: "GENERAL",
  CARDIOLOGY: "CARDIOLOGY",
  PULMONOLOGY: "PULMONOLOGY",
  NEUROLOGY: "NEUROLOGY",
  INTERNAL_MEDICINE: "INTERNAL_MEDICINE",
  EMERGENCY_MEDICINE: "EMERGENCY_MEDICINE",
  PEDIATRICS: "PEDIATRICS",
  ORTHOPEDICS: "ORTHOPEDICS",
  DERMATOLOGY: "DERMATOLOGY",
  PSYCHIATRY: "PSYCHIATRY",
} as const;

export type CreateCaseRequestGender =
  (typeof CreateCaseRequestGender)[keyof typeof CreateCaseRequestGender];

export const CreateCaseRequestGender = {
  MALE: "MALE",
  FEMALE: "FEMALE",
  OTHER: "OTHER",
} as const;

export type CreateCaseRequestReportType =
  (typeof CreateCaseRequestReportType)[keyof typeof CreateCaseRequestReportType];

export const CreateCaseRequestReportType = {
  SOAP: "SOAP",
  JSON: "JSON",
} as const;

export interface VitalSigns {
  temperature?: number;
  heartRate?: number;
  bloodPressureSystolic?: number;
  bloodPressureDiastolic?: number;
  oxygenSaturation?: number;
  respiratoryRate?: number;
}

export interface CreateCaseRequest {
  patientName: string;
  age: number;
  gender: CreateCaseRequestGender;
  chiefComplaint: string;
  symptoms: string[];
  vitalSigns?: VitalSigns;
  medicalHistory?: string;
  currentMedications?: string;
  reportType: CreateCaseRequestReportType;
  rawReport?: string;
}

export interface AssignDoctorRequest {
  doctorId: number;
}

export type DiagnosisRequestStatus =
  (typeof DiagnosisRequestStatus)[keyof typeof DiagnosisRequestStatus];

export const DiagnosisRequestStatus = {
  DIAGNOSIS_IN_PROGRESS: "DIAGNOSIS_IN_PROGRESS",
  COMPLETED: "COMPLETED",
} as const;

export interface DiagnosisRequest {
  diagnosisNotes: string;
  status?: DiagnosisRequestStatus;
}

export interface UpdateStatusRequest {
  caseStatus: CaseStatus;
}

export interface ClinicalCase {
  id: number;
  patientName: string;
  age: number;
  gender: string;
  chiefComplaint: string;
  symptoms: string[];
  vitalSigns?: VitalSigns;
  medicalHistory?: string;
  currentMedications?: string;
  reportType: string;
  rawReport?: string;
  riskLevel: RiskLevel;
  recommendedAction: RecommendedAction;
  riskFactors: string[];
  briefEnglish: string;
  briefArabic: string;
  acknowledged: boolean;
  caseStatus: CaseStatus;
  assignedDoctorId?: number;
  diagnosisNotes?: string;
  recommendedDepartment?: string;
  createdAt: string;
  updatedAt: string;
}

export interface Hospital {
  id: number;
  nameEn: string;
  nameAr: string;
  city: string;
  address: string;
  phone: string;
  specialties: string[];
  isAvailable: string;
  level: string;
}

export type CreateTransferRequestDirection =
  (typeof CreateTransferRequestDirection)[keyof typeof CreateTransferRequestDirection];

export const CreateTransferRequestDirection = {
  OUTGOING: "OUTGOING",
  INCOMING: "INCOMING",
} as const;

export type CreateTransferRequestTransportMethod =
  (typeof CreateTransferRequestTransportMethod)[keyof typeof CreateTransferRequestTransportMethod];

export const CreateTransferRequestTransportMethod = {
  AMBULANCE: "AMBULANCE",
  AIR_AMBULANCE: "AIR_AMBULANCE",
  PRIVATE_VEHICLE: "PRIVATE_VEHICLE",
  HOSPITAL_TRANSPORT: "HOSPITAL_TRANSPORT",
} as const;

export interface CreateTransferRequest {
  caseId?: number;
  direction: CreateTransferRequestDirection;
  toHospitalId?: number;
  fromHospitalId?: number;
  patientName: string;
  patientAge: number;
  patientGender: string;
  chiefComplaint: string;
  riskLevel: string;
  transportMethod: CreateTransferRequestTransportMethod;
  clinicalSummary: string;
  specialRequirements?: string;
  estimatedArrival?: string;
}

export interface Transfer {
  id: number;
  caseId?: number;
  direction: string;
  fromHospitalId: number;
  toHospitalId: number;
  fromHospital?: Hospital;
  toHospital?: Hospital;
  patientName: string;
  patientAge: number;
  patientGender: string;
  chiefComplaint: string;
  riskLevel: string;
  transportMethod: string;
  clinicalSummary: string;
  specialRequirements?: string;
  estimatedArrival?: string;
  status: TransferStatus;
  rejectionReason?: string;
  transferCode: string;
  createdAt: string;
  updatedAt: string;
}

export interface Doctor {
  id: number;
  nameEn: string;
  nameAr: string;
  specialty: Specialty;
  specialtyLabelEn: string;
  specialtyLabelAr: string;
  qualifications: string;
  /** Years of experience */
  experience: number;
  rating: number;
  avatarInitials: string;
  isAvailable: boolean;
  consultationFeeUsd: number;
  languages: string[];
}

export type CreateConsultationRequestPreferredLanguage =
  (typeof CreateConsultationRequestPreferredLanguage)[keyof typeof CreateConsultationRequestPreferredLanguage];

export const CreateConsultationRequestPreferredLanguage = {
  EN: "EN",
  AR: "AR",
} as const;

export interface CreateConsultationRequest {
  caseId: number;
  doctorId: number;
  scheduledAt: string;
  patientNotes?: string;
  preferredLanguage?: CreateConsultationRequestPreferredLanguage;
}

export interface Consultation {
  id: number;
  caseId: number;
  doctorId: number;
  doctor?: Doctor;
  patientName: string;
  scheduledAt: string;
  status: ConsultationStatus;
  patientNotes?: string;
  preferredLanguage?: string;
  meetingRoomId: string;
  createdAt: string;
}

export interface JoinConsultationResponse {
  consultationId: number;
  meetingRoomId: string;
  joinUrl: string;
  status: ConsultationStatus;
}

export type RejectTransferBody = {
  reason?: string;
};