# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all clinical cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClinicalCase"
    post:
      operationId: createCase
      tags: [cases]
      summary: Submit a new clinical case report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Case created and analyzed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalCase"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a specific case by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalCase"
        "404":
          description: Case not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /cases/{id}/acknowledge:
    post:
      operationId: acknowledgeCase
      tags: [cases]
      summary: Acknowledge a CRITICAL case alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalCase"
  /cases/{id}/assign-doctor:
    post:
      operationId: assignDoctor
      tags: [cases]
      summary: Assign a doctor to a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDoctorRequest"
      responses:
        "200":
          description: Doctor assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalCase"
  /cases/{id}/diagnosis:
    post:
      operationId: updateDiagnosis
      tags: [cases]
      summary: Doctor records diagnosis notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisRequest"
      responses:
        "200":
          description: Diagnosis updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalCase"
  /cases/{id}/update-status:
    post:
      operationId: updateCaseStatus
      tags: [cases]
      summary: Update the lifecycle status of a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalCase"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    RiskLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    CaseStatus:
      type: string
      enum:
        - REPORT_RECEIVED
        - ANALYZED
        - CASE_CREATED
        - ROUTED_TO_HOSPITAL
        - RECEIVED_BY_HOSPITAL
        - ASSIGNED_TO_DOCTOR
        - DIAGNOSIS_IN_PROGRESS
        - COMPLETED
    RecommendedAction:
      type: string
      enum:
        - SELF_CARE
        - SCHEDULE_CONSULTATION
        - HOSPITAL_VISIT
        - EMERGENCY_RESPONSE
...
    CreateCaseRequest:
      type: object
      properties:
        patientName:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        chiefComplaint:
          type: string
        symptoms:
          type: array
          items:
            type: string
        vitalSigns:
          $ref: "#/components/schemas/VitalSigns"
        medicalHistory:
          type: string
        currentMedications:
          type: string
        reportType:
          type: string
          enum: [SOAP, JSON]
        rawReport:
          type: string
      required:
        - patientName
        - age
        - gender
        - chiefComplaint
        - symptoms
        - reportType
    AssignDoctorRequest:
      type: object
      properties:
        doctorId:
          type: integer
      required:
        - doctorId
    DiagnosisRequest:
      type: object
      properties:
        diagnosisNotes:
          type: string
        status:
          type: string
          enum: [DIAGNOSIS_IN_PROGRESS, COMPLETED]
      required:
        - diagnosisNotes
    UpdateStatusRequest:
      type: object
      properties:
        caseStatus:
          $ref: "#/components/schemas/CaseStatus"
      required:
        - caseStatus
    VitalSigns:
      type: object
      properties:
        temperature:
          type: number
        heartRate:
          type: integer
        bloodPressureSystolic:
          type: integer
        bloodPressureDiastolic:
          type: integer
        oxygenSaturation:
          type: number
        respiratoryRate:
          type: integer
    ClinicalCase:
      type: object
      properties:
        id:
          type: integer
        patientName:
          type: string
        age:
          type: integer
        gender:
          type: string
        chiefComplaint:
          type: string
        symptoms:
          type: array
          items:
            type: string
        vitalSigns:
          $ref: "#/components/schemas/VitalSigns"
        medicalHistory:
          type: string
        currentMedications:
          type: string
        reportType:
          type: string
        rawReport:
          type: string
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        recommendedAction:
          $ref: "#/components/schemas/RecommendedAction"
        riskFactors:
          type: array
          items:
            type: string
        briefEnglish:
          type: string
        briefArabic:
          type: string
        acknowledged:
          type: boolean
        caseStatus:
          $ref: "#/components/schemas/CaseStatus"
        assignedDoctorId:
          type: integer
        diagnosisNotes:
          type: string
        recommendedDepartment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - patientName
        - age
        - gender
        - chiefComplaint
        - symptoms
        - reportType
        - riskLevel
        - recommendedAction
        - riskFactors
        - briefEnglish
        - briefArabic
        - acknowledged
        - caseStatus
        - createdAt
        - updatedAt
...