# lib/api-zod/src/generated/types/assignDoctorRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignDoctorRequest {
  doctorId: number;
}