# lib/api-zod/src/generated/types/caseStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  REPORT_RECEIVED: "REPORT_RECEIVED",
  ANALYZED: "ANALYZED",
  CASE_CREATED: "CASE_CREATED",
  ROUTED_TO_HOSPITAL: "ROUTED_TO_HOSPITAL",
  RECEIVED_BY_HOSPITAL: "RECEIVED_BY_HOSPITAL",
  ASSIGNED_TO_DOCTOR: "ASSIGNED_TO_DOCTOR",
  DIAGNOSIS_IN_PROGRESS: "DIAGNOSIS_IN_PROGRESS",
  COMPLETED: "COMPLETED",
} as const;