# lib/api-zod/src/generated/types/clinicalCase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStatus } from "./caseStatus";
import type { RecommendedAction } from "./recommendedAction";
import type { RiskLevel } from "./riskLevel";
import type { VitalSigns } from "./vitalSigns";

export interface ClinicalCase {
  id: number;
  patientName: string;
  age: number;
  gender: string;
  chiefComplaint: string;
  symptoms: string[];
  vitalSigns?: VitalSigns;
  medicalHistory?: string;
  currentMedications?: string;
  reportType: string;
  rawReport?: string;
  riskLevel: RiskLevel;
  recommendedAction: RecommendedAction;
  riskFactors: string[];
  briefEnglish: string;
  briefArabic: string;
  acknowledged: boolean;
  caseStatus: CaseStatus;
  assignedDoctorId?: number;
  diagnosisNotes?: string;
  recommendedDepartment?: string;
  createdAt: Date;
  updatedAt: Date;
}