# lib/api-zod/src/generated/types/diagnosisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosisRequestStatus } from "./diagnosisRequestStatus";

export interface DiagnosisRequest {
  diagnosisNotes: string;
  status?: DiagnosisRequestStatus;
}